remodeling - selectivly seperate style source from component source
