+++ src/styles/StyledRankingAnyTagItem.tsx
@@ -4,7 +4,7 @@ interface AnyTagBtnProps {
   isSelected: boolean;
 }
 const StyledRankingAnyTagItem = styled.button<AnyTagBtnProps>`
-  color: ${({ theme, isSelected }) => (isSelected ? theme.palette.blue800 : theme.palette.blue600)};
+  color: ${({ theme, isSelected }) => (isSelected ? theme.palette.blue700 : theme.palette.blue400)};
   ${({ theme, isSelected }) => (isSelected ? theme.typography.label1Bold : theme.typography.label1Regular)}
   border: none;
   background-color: inherit;
