--- src/styles/StyledRankingSexTagItem.tsx
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-const StyledRankingSexTagItem = styled.div`
-  width: 60px;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 10px;
-`;
-
-export default StyledRankingSexTagItem;
