+++ src/styles/globalStyles.css
@@ -14,3 +14,10 @@ body {
   min-width: 320px;
   width: 100%;
 }
+
+@font-face {
+  font-family: 'Pretendard-Regular';
+  src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
+  font-weight: 400;
+  font-style: normal;
+}
