+++ src/styles/reset.ts
@@ -1,15 +1,7 @@
 // /src/styles/global.ts
-import { css } from '@emotion/react'
+import { css } from '@emotion/react';
 
 export const reset = css`
-  @font-face {
-    font-family: 'Pretendard-Regular';
-    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff')
-      format('woff');
-    font-weight: 400;
-    font-style: normal;
-  }
-
   html,
   body,
   div,
@@ -135,7 +127,6 @@ export const reset = css`
   }
 
   * {
-    font-family:
-      'Pretendard-Regular', 'Apple SD Gothic Neo', 'Malgun Gothic', sans-serif, 'Noto Sans KR';
+    font-family: 'Pretendard-Regular', 'Apple SD Gothic Neo', 'Malgun Gothic', sans-serif, 'Noto Sans KR';
   }
-`
+`;
